Fixed some issues with errorhandling
Hopefully the fileformat is better now. 

Changed the code in the init section. Needed to use global variables.
Fixed issue then a wrong address in in the database caused cbpi in whole stop working. Only writing in the database can solve this. Delete database is not working when the default adress is wrong.
Changed LCD_Adress-->LCD_Address in parameters
Better bierglass icon ;-)